warehouse: refactor security policies
Punt principal handling further down, remove the generic
identity implementation, etc. etc.
