+++ warehouse/accounts/__init__.py
@@ -10,18 +10,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import datetime
-import enum
-
 from pyramid.authorization import ACLAuthorizationPolicy
-from pyramid.httpexceptions import HTTPUnauthorized
 
 from warehouse.accounts.interfaces import (
     IPasswordBreachedService,
     ITokenService,
     IUserService,
 )
-from warehouse.accounts.models import DisableReason
 from warehouse.accounts.security_policy import (
     BasicAuthSecurityPolicy,
     SessionSecurityPolicy,
@@ -32,12 +27,6 @@ from warehouse.accounts.services import (
     NullPasswordBreachedService,
     TokenServiceFactory,
     database_login_factory,
-)
-from warehouse.email import send_password_compromised_email_hibp
-from warehouse.errors import (
-    BasicAuthAccountFrozen,
-    BasicAuthBreachedPassword,
-    BasicAuthFailedPassword,
 )
 from warehouse.macaroons.security_policy import (
     MacaroonAuthorizationPolicy,
@@ -50,125 +39,6 @@ __all__ = ["NullPasswordBreachedService", "HaveIBeenPwnedPasswordBreachedService
 
 
 REDIRECT_FIELD_NAME = "next"
-
-
-class AuthenticationMethod(enum.Enum):
-    BASIC_AUTH = "basic-auth"
-    SESSION = "session"
-    MACAROON = "macaroon"
-
-
-def _format_exc_status(exc, message):
-    exc.status = f"{exc.status_code} {message}"
-    return exc
-
-
-def _authenticate(userid, request):
-    """Apply the necessary principals to the authenticated user"""
-    login_service = request.find_service(IUserService, context=None)
-    user = login_service.get_user(userid)
-
-    if user is None:
-        return
-
-    if request.session.password_outdated(login_service.get_password_timestamp(userid)):
-        request.session.invalidate()
-        request.session.flash("Session invalidated by password change", queue="error")
-        return
-
-    principals = []
-
-    if user.is_superuser:
-        principals.append("group:admins")
-    if user.is_moderator or user.is_superuser:
-        principals.append("group:moderators")
-    if user.is_psf_staff or user.is_superuser:
-        principals.append("group:psf_staff")
-
-    # user must have base admin access if any admin permission
-    if principals:
-        principals.append("group:with_admin_dashboard_access")
-
-    return principals
-
-
-def _basic_auth_check(username, password, request):
-    request.authentication_method = AuthenticationMethod.BASIC_AUTH
-
-    # Basic authentication can only be used for uploading
-    if request.matched_route.name not in ["forklift.legacy.file_upload"]:
-        return
-
-    login_service = request.find_service(IUserService, context=None)
-    breach_service = request.find_service(IPasswordBreachedService, context=None)
-
-    userid = login_service.find_userid(username)
-    if userid is not None:
-        user = login_service.get_user(userid)
-        is_disabled, disabled_for = login_service.is_disabled(user.id)
-        if is_disabled:
-            # This technically violates the contract a little bit, this function is
-            # meant to return None if the user cannot log in. However we want to present
-            # a different error message than is normal when we're denying the log in
-            # because of a compromised password. So to do that, we'll need to raise a
-            # HTTPError that'll ultimately get returned to the client. This is OK to do
-            # here because we've already successfully authenticated the credentials, so
-            # it won't screw up the fall through to other authentication mechanisms
-            # (since we wouldn't have fell through to them anyways).
-            if disabled_for == DisableReason.CompromisedPassword:
-                raise _format_exc_status(
-                    BasicAuthBreachedPassword(), breach_service.failure_message_plain
-                )
-            elif disabled_for == DisableReason.AccountFrozen:
-                raise _format_exc_status(BasicAuthAccountFrozen(), "Account is frozen.")
-            else:
-                raise _format_exc_status(HTTPUnauthorized(), "Account is disabled.")
-        elif login_service.check_password(
-            user.id,
-            password,
-            tags=["mechanism:basic_auth", "method:auth", "auth_method:basic"],
-        ):
-            if breach_service.check_password(
-                password, tags=["method:auth", "auth_method:basic"]
-            ):
-                send_password_compromised_email_hibp(request, user)
-                login_service.disable_password(
-                    user.id, reason=DisableReason.CompromisedPassword
-                )
-                raise _format_exc_status(
-                    BasicAuthBreachedPassword(), breach_service.failure_message_plain
-                )
-
-            login_service.update_user(user.id, last_login=datetime.datetime.utcnow())
-            return _authenticate(user.id, request)
-        else:
-            user.record_event(
-                tag="account:login:failure",
-                ip_address=request.remote_addr,
-                additional={"reason": "invalid_password", "auth_method": "basic"},
-            )
-            raise _format_exc_status(
-                BasicAuthFailedPassword(),
-                "Invalid or non-existent authentication information. "
-                "See {projecthelp} for more information.".format(
-                    projecthelp=request.help_url(_anchor="invalid-auth")
-                ),
-            )
-
-
-def _session_authenticate(userid, request):
-    request.authentication_method = AuthenticationMethod.SESSION
-
-    # Session authentication cannot be used for uploading
-    if request.matched_route.name in ["forklift.legacy.file_upload"]:
-        return
-
-    return _authenticate(userid, request)
-
-
-def _macaroon_authenticate(userid, request):
-    request.authentication_method = AuthenticationMethod.MACAROON
-    return _authenticate(userid, request)
 
 
 def _user(request):
@@ -213,9 +83,9 @@ def includeme(config):
     config.set_security_policy(
         MultiSecurityPolicy(
             [
-                SessionSecurityPolicy(callback=_session_authenticate),
-                BasicAuthSecurityPolicy(check=_basic_auth_check),
-                MacaroonSecurityPolicy(callback=_macaroon_authenticate),
+                SessionSecurityPolicy(),
+                BasicAuthSecurityPolicy(),
+                MacaroonSecurityPolicy(),
             ],
             authz_policy,
         )
