+++ warehouse/accounts/security_policy.py
@@ -10,49 +10,156 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import datetime
+
 from pyramid.authentication import (
     SessionAuthenticationHelper,
     extract_http_basic_credentials,
 )
+from pyramid.httpexceptions import HTTPUnauthorized
 from pyramid.interfaces import IAuthorizationPolicy, ISecurityPolicy
 from pyramid.threadlocal import get_current_request
 from zope.interface import implementer
 
-from warehouse.accounts.interfaces import IUserService
+from warehouse.accounts.interfaces import IPasswordBreachedService, IUserService
+from warehouse.accounts.models import DisableReason
 from warehouse.cache.http import add_vary_callback
-from warehouse.errors import WarehouseDenied
+from warehouse.email import send_password_compromised_email_hibp
+from warehouse.errors import (
+    BasicAuthAccountFrozen,
+    BasicAuthBreachedPassword,
+    BasicAuthFailedPassword,
+    WarehouseDenied,
+)
 from warehouse.packaging.models import TwoFactorRequireable
-from warehouse.utils.security_policy import SecurityPolicy
+from warehouse.utils.security_policy import AuthenticationMethod
+
+
+def _format_exc_status(exc, message):
+    exc.status = f"{exc.status_code} {message}"
+    return exc
+
+
+def _basic_auth_check(username, password, request):
+    # Basic authentication can only be used for uploading
+    if request.matched_route.name not in ["forklift.legacy.file_upload"]:
+        return False
+
+    login_service = request.find_service(IUserService, context=None)
+    breach_service = request.find_service(IPasswordBreachedService, context=None)
+
+    userid = login_service.find_userid(username)
+    if userid is not None:
+        user = login_service.get_user(userid)
+        is_disabled, disabled_for = login_service.is_disabled(user.id)
+        if is_disabled:
+            # This technically violates the contract a little bit, this function is
+            # meant to return False if the user cannot log in. However we want to present
+            # a different error message than is normal when we're denying the log in
+            # because of a compromised password. So to do that, we'll need to raise a
+            # HTTPError that'll ultimately get returned to the client. This is OK to do
+            # here because we've already successfully authenticated the credentials, so
+            # it won't screw up the fall through to other authentication mechanisms
+            # (since we wouldn't have fell through to them anyways).
+            if disabled_for == DisableReason.CompromisedPassword:
+                raise _format_exc_status(
+                    BasicAuthBreachedPassword(), breach_service.failure_message_plain
+                )
+            elif disabled_for == DisableReason.AccountFrozen:
+                raise _format_exc_status(BasicAuthAccountFrozen(), "Account is frozen.")
+            else:
+                raise _format_exc_status(HTTPUnauthorized(), "Account is disabled.")
+        elif login_service.check_password(
+            user.id,
+            password,
+            tags=["mechanism:basic_auth", "method:auth", "auth_method:basic"],
+        ):
+            if breach_service.check_password(
+                password, tags=["method:auth", "auth_method:basic"]
+            ):
+                send_password_compromised_email_hibp(request, user)
+                login_service.disable_password(
+                    user.id, reason=DisableReason.CompromisedPassword
+                )
+                raise _format_exc_status(
+                    BasicAuthBreachedPassword(), breach_service.failure_message_plain
+                )
+
+            login_service.update_user(user.id, last_login=datetime.datetime.utcnow())
+            return True
+        else:
+            user.record_event(
+                tag="account:login:failure",
+                ip_address=request.remote_addr,
+                additional={"reason": "invalid_password", "auth_method": "basic"},
+            )
+            raise _format_exc_status(
+                BasicAuthFailedPassword(),
+                "Invalid or non-existent authentication information. "
+                "See {projecthelp} for more information.".format(
+                    projecthelp=request.help_url(_anchor="invalid-auth")
+                ),
+            )
 
 
 @implementer(ISecurityPolicy)
-class BasicAuthSecurityPolicy(SecurityPolicy):
-    def __init__(self, check):
-        super().__init__(self.callback)
+class SessionSecurityPolicy:
+    def __init__(self):
+        self._session_helper = SessionAuthenticationHelper()
+        self._principals = []
 
-        self._check = check
+    def identity(self, request):
+        # If we're calling into this API on a request, then we want to register
+        # a callback which will ensure that the response varies based on the
+        # Cookie header.
+        request.add_response_callback(add_vary_callback("Cookie"))
+        request.authentication_method = AuthenticationMethod.SESSION
 
+        # NOTE: Why is this authenticated_userid? I guess it's already
+        # been validated by session validation?
+        userid = self._session_helper.authenticated_userid(request)
+
+        # Session authentication cannot be used for uploading
+        if request.matched_route.name in ["forklift.legacy.file_upload"]:
+            return None
+
+        # Sessions can only authenticate users, not any other type of identity.
+        login_service = request.find_service(IUserService, context=None)
+        return login_service.get_user(userid)
+
+    def forget(self, request, **kw):
+        return self._session_helper.forget(request, **kw)
+
+    def remember(self, request, userid, **kw):
+        return self._session_helper.remember(request, userid, **kw)
+
+
+@implementer(ISecurityPolicy)
+class BasicAuthSecurityPolicy:
     def callback(self, userid, request):
         if credentials := extract_http_basic_credentials(request):
             username, password = credentials
             return self._check(username, password, request)
         return None
 
-    def unauthenticated_userid(self, request):
+    def identity(self, request):
         # If we're calling into this API on a request, then we want to register
         # a callback which will ensure that the response varies based on the
         # Authorization header.
         request.add_response_callback(add_vary_callback("Authorization"))
+        request.authentication_method = AuthenticationMethod.BASIC_AUTH
 
         credentials = extract_http_basic_credentials(request)
         if credentials is None:
             return None
 
-        login_service = request.find_service(IUserService, context=None)
-        if userid := login_service.find_userid(credentials.username):
-            return str(userid)
+        username, password = credentials
+        if not _basic_auth_check(username, password, request):
+            return None
 
-        return None
+        # Like sessions; basic auth can only authenticate users.
+        login_service = request.find_service(IUserService, context=None)
+        return login_service.get_user_by_username(username)
 
     def forget(self, request, **kw):
         # No-op.
@@ -61,27 +168,6 @@ class BasicAuthSecurityPolicy(SecurityPolicy):
     def remember(self, request, userid, **kw):
         # NOTE: We could make realm configurable here.
         return [("WWW-Authenticate", 'Basic realm="Realm"')]
-
-
-@implementer(ISecurityPolicy)
-class SessionSecurityPolicy(SecurityPolicy):
-    def __init__(self, callback=None):
-        super().__init__(callback)
-        self._session_helper = SessionAuthenticationHelper()
-
-    def unauthenticated_userid(self, request):
-        # If we're calling into this API on a request, then we want to register
-        # a callback which will ensure that the response varies based on the
-        # Cookie header.
-        request.add_response_callback(add_vary_callback("Cookie"))
-
-        return self._session_helper.authenticated_userid(request)
-
-    def forget(self, request, **kw):
-        return self._session_helper.forget(request, **kw)
-
-    def remember(self, request, userid, **kw):
-        return self._session_helper.remember(request, userid, **kw)
 
 
 @implementer(IAuthorizationPolicy)
